chore: cleanup engine builder
