+++ src/engine/engine.ts
@@ -22,20 +22,7 @@
  *
  * @copyright (C) lemoncloud.io 2019 - All Rights Reserved.
  */
-//! create engine in global scope (WARN! should run top level)
-import engine, { LemonEngine } from '../engine/index';
-export const $engine: LemonEngine = engine(global, { env: process.env });
-
-//! re-use core modules.
-export const $U = $engine.U;
-export const $_ = $engine._;
-if (!$U) throw new Error('$U(utilities) is required!');
-if (!$_) throw new Error('$_(lodash) is required!');
-
-//! export common(log) functions
-export const _log = $engine.log;
-export const _inf = $engine.inf;
-export const _err = $engine.err;
+import { $engine, $U, _log, _inf, _err } from './index';
 
 //! import sub-modules.
 import { SlackPostBody, MetricPostBody } from '../common/types';
